GetLabels: also search included files!
Also, SetMainFile: when the main .tex file is not found, do not throw an
error. It might be the case that the user is just beginning, and thus
there is no main file, yet... An error should be thrown only when the
user invokes some functionality that requires the main file, but one
cannot be found.