Added a new function SetMainFileOrThrowUp().
The ideia is that this function calls SetMainFile, and throws an
exception if a main file could not be found.

SetMainFile does nothing if the main cannot be found, because it is
called whenever a .tex file is opened -- and at this point, there might
not be a main file yet.