Change imod mappings, and math completion.
First regarding math completion, as it was it did not function when
calling with nothing written (a:base empty). Now it works.

Next, I reorganised the imode mappings. In particular, those of the form
\varsomething now have the map <localleader>_<key>, where
<localleader><key> is the mapping for \something.

Also deleted the map for :copen, as it is no longer used.

And added a map <localleader><, with produces \langle \rangle