Main addition: preview of graphics files.
Now, in addition to being able to preview labels and \cites, you can
can also preview graphics files. If the graphic file is pdf, then the
pdf viewer is used. Otherwise (jpg/png), it requires for an image viewer
to be set in config, using the key viewer_images.

Fixed a bug: when viewing pdf (<Localleader>V), the function now checks
that there is a pdf viewwe defined.

gp/gd now also works in \includeonly, in the same manner it works for
\include and \input.