Fixed two bugs in env selection operators
Firstly, when selecting the inner environment, if the cursor was placed
at the \end line, that line was also included in the visual selection,
which was not supposed to be the case. That is now fixed (it entailed
recoding function EnvironmentOperator()).

Also, selection did not work with two or more nested environments, eg:

    \section{whatever}
      foo

      \begin{foo}
        blah <cursor here>
        \begin{bar}
          \begin{baz}
            sdf
          \end{baz}
        \end{bar}
        sdf
      \end{foo}

      sdfsdf

That has now been fixed (functions FindEndBelow() and FindBeginAbove()).