replace cmd arg visual selection with regexp
for say, \input{foo}, the selection of foo was done with a visual
selection, that got yanked. This cleared any thing the user himself had
previously yanked (and presumably wanted to paste in file foo).

Said selection is now done with regexp, thus preserving the user's
previous yanked/deleted/etc text, if any.