Improve math completion.
Now, the math completion dictionary uses the "info" entry to store
concepts related to the symbol. For example, both \cap and \cup are
related to sets, and the first to intersection, and the second to union.

This new information can now also be used for completion. E.g., typying
union^X^O will show, in the completion popup, also \cap, even though
union is not a part of the command name.