Still the scope of config vars...
The default config, in ftplugin/tex.vim, is a script var:
s:tex_seven_config. The global config, g:tex_seven_config, is what is
used throughout the plugin.

Also, in ftplugin/tex.vim, we now reset the g:maplocalleader to its
previous value, if any, at the end of the file (if there was no previous
value, it is unset).