Allow for auto-completion beyond auto-completion
For example, typing <Leader>A inserts \includeonly{} and triggers
auto-completion. Before this commit, the text inserted inside the curly
brackets was matched against THE START of completion possibilities (and
case sensitive). Now it is case insensitive, and the inserted text is
matched at any position (not necessarily at the start).