Added possibility to choose how \label retrieval is done.
A new variable was added to the config dict: label_retrieval_use_script.
By default, it is 0, which means \label retrieval is done using
vimscript only. If the user sets it to 1, then \label retrieval is done
using the perl script in the scripts/ folder. This requires perl
installed, together with the perl-json module.